+++ kernel/src/arch/x86_64/interrupts/IRQ.cpp
@@ -14,7 +14,7 @@ void x86_64_IRQ_Handler(x86_64_Interrupt_Registers* regs) {
         g_IRQHandlers[irq](regs);
     }
     else {
-        printf("Unhandled IRQ %d...\n", irq);
+        fprintf(VFS_DEBUG, "Unhandled IRQ %d...\n", irq);
     }
 
     x86_64_PIC_sendEOI(irq);