+++ CHANGES.md
@@ -1,6 +1,19 @@
 # Changes
 
-## Latest Changes - 28/06/2023
+## Latest Changes - 30/06/2023
+
+- Added Process and Thread classes
+- Added x86_64 task switch code
+- Added x86_64 register save code
+- Added `x86_64_get_stack_ptr` function
+- Added utility function to get a new stack (for tasks)
+- Added `x86_64_Interrupt_Registers` and `x86_64_Registers` conversions
+- Implemented basic kernel round-robin scheduler
+- Added function call from `x86_64_PIT_Handler` to `Scheduling::Scheduler::TimerTick()`
+- Changed kernel init to a 2 stage initialisation with second stage being called by the scheduler
+- Renamed `LATE_STAGE` to `STAGE2`
+
+## 28/06/2023
 
 - Fixed a redefinition error in `HAL/time.cpp`
 - Removed unnecessary includes and function calls from `StartKernel`